MAILBOX-320 Improve mailboxManager testing speed
On some backend, creating 81 mailbox and 243 messages sequantially can be "slow"
